+++ addon/components/colsconfig-dialog-content.js
@@ -1,6 +1,8 @@
 import Ember from 'ember';
 import FlexberryBaseComponent from './flexberry-base-component';
 import serializeSortingParam from '../utils/serialize-sorting-param';
+import QueryBuilder from 'ember-flexberry-data/query/builder';
+import ODataAdapter from 'ember-flexberry-data/query/odata-adapter';
 const { getOwner } = Ember;
 const _idPrefix = 'ColDesc';
 
@@ -19,6 +21,15 @@ export default FlexberryBaseComponent.extend({
    @default Ember.inject.service()
    */
   colsConfigMenu: Ember.inject.service(),
+
+  /**
+   Service that triggers objectlistview events.
+
+   @property objectlistviewEventsService
+   @type {Class}
+   @default Ember.inject.service()
+   */
+  objectlistviewEventsService: Ember.inject.service('objectlistview-events'),
 
   /**
    Model with added DOM elements.
@@ -74,6 +85,33 @@ export default FlexberryBaseComponent.extend({
   */
   perPageValue: undefined,
 
+  /**
+    Params for export excel.
+
+    @property exportParams
+    @type {Object}
+    @default {}
+  */
+  exportParams: {},
+
+  /**
+    Current store.
+
+    @property store
+    @type {Object}
+    @default undefined
+  */
+  store: undefined,
+
+  /**
+    Current model name.
+
+    @property modelName
+    @type {String}
+    @default undefined
+  */
+  modelName: undefined,
+
   init: function() {
     this._super(...arguments);
     this.modelForDOM = [];
@@ -85,6 +123,9 @@ export default FlexberryBaseComponent.extend({
     this.componentName = this.model.componentName;
     this.perPageValue = this.model.perPageValue;
     this.saveColWidthState = this.model.saveColWidthState;
+    this.exportParams = this.model.exportParams;
+    this.modelName = this.model.modelName;
+    this.set('store', this.model.store);
     let colDescs = this.model.colDescs;
     for (let i = 0; i < colDescs.length; i++) {
       let colDesc = colDescs[i];
@@ -115,6 +156,15 @@ export default FlexberryBaseComponent.extend({
     firstButtonUp.addClass('disabled'); // Disable first button up
     let lastButtondown = Ember.$('#ColDescRowDown_' + (this.modelForDOM.length - 1));
     lastButtondown.addClass('disabled'); // Disable last button down
+    if (this.exportParams.isExportExcel && this.exportParams.immediateExport) {
+      this.exportParams.immediateExport = false;
+      this.actions.apply.call(this);
+    }
+  },
+
+  didInsertElement: function() {
+    this._super(...arguments);
+    this.$('.sort-direction-dropdown').dropdown();
   },
 
   actions: {
@@ -311,22 +361,54 @@ export default FlexberryBaseComponent.extend({
      @method actions.apply
     */
     apply: function() {
-      let colsConfig = this._getSettings();
-      let settingName =  Ember.$('#columnConfigurtionSettingName')[0].value.trim();
-      if (settingName.length > 0 && this._isChanged && !confirm(this.get('i18n').t('components.colsconfig-dialog-content.use-without-save') + settingName)) {
-        return;
-      }
-
-      //Save colsConfig in userSettings as DEFAULT
-      let router = getOwner(this).lookup('router:main');
-      let savePromise = this._getSavePromise(undefined, colsConfig);
-      savePromise.then(
-        record => {
-          let sort = serializeSortingParam(colsConfig.sorting);
-          router.router.transitionTo(router.currentRouteName, { queryParams: { sort: sort, perPage: colsConfig.perPage || 5 } });
+      if (!this.exportParams.isExportExcel) {
+        let colsConfig = this._getSettings();
+        let settingName =  Ember.$('#columnConfigurtionSettingName')[0].value.trim();
+        if (settingName.length > 0 && this._isChanged && !confirm(this.get('i18n').t('components.colsconfig-dialog-content.use-without-save') + settingName)) {
+          return;
         }
-      );
-      this.sendAction('close', colsConfig); // close modal window
+
+        //Save colsConfig in userSettings as DEFAULT
+        let router = getOwner(this).lookup('router:main');
+        let savePromise = this._getSavePromise(undefined, colsConfig);
+        savePromise.then(
+          record => {
+            let sort = serializeSortingParam(colsConfig.sorting);
+            router.router.transitionTo(router.currentRouteName, { queryParams: { sort: sort, perPage: colsConfig.perPage || 5 } });
+          }
+        );
+        this.sendAction('close', colsConfig); // close modal window
+      } else {
+        let _this = this;
+        _this.get('objectlistviewEventsService').setLoadingState('loading');
+        let store = this.get('store.onlineStore') || this.get('store');
+        let adapter = store.adapterFor(this.modelName);
+        let currentQuery = this._getCurrentQuery();
+        adapter.query(store, this.modelName, currentQuery).then((result) => {
+          let blob = new Blob([result], { type: 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet' });
+          let anchor = Ember.$('.download-anchor');
+          if (!Ember.isBlank(anchor)) {
+            if (window.navigator.msSaveOrOpenBlob) {
+              let downloadFunction = function() {
+                window.navigator.msSaveOrOpenBlob(blob, 'list.xlsx');
+              };
+
+              anchor.on('click', downloadFunction);
+              anchor.get(0).click();
+              anchor.off('click', downloadFunction);
+            } else {
+              let downloadUrl = URL.createObjectURL(blob);
+              anchor.prop('href', downloadUrl);
+              anchor.prop('download', 'list.xlsx');
+              anchor.get(0).click();
+            }
+          }
+
+          this.get('objectlistviewEventsService').setLoadingState('');
+        }).catch(() => {
+          this.get('objectlistviewEventsService').setLoadingState('');
+        });
+      }
     },
     /**
       Save named settings specified in the interface as named values
@@ -392,10 +474,69 @@ export default FlexberryBaseComponent.extend({
     perPageChanged: function() {
       this._changed();
     },
+
+    /**
+      DetSeparateCols value is changed
+
+      @method actions.detSeparateColsChange
+    */
+    detSeparateColsChange: function() {
+      this._changed();
+    },
+
+    /**
+      DetSeparateRows value is changed
+
+      @method actions.detSeparateRowsChange
+    */
+    detSeparateRowsChange: function() {
+      this._changed();
+    },
+  },
+
+  /**
+    Gets current query for export excel
+
+    @method _getCurrentQuery
+  */
+  _getCurrentQuery: function() {
+    let settings = this._getSettings();
+    let sortString = '';
+    let modelName = this.modelName;
+    settings.sorting.map(sort => {
+      sortString += `${sort.propName} ${sort.direction},`;
+    });
+    sortString = sortString.slice(0, -1);
+    let store = this.get('store.onlineStore') || this.get('store');
+    let builder = new QueryBuilder(store, modelName);
+    let adapter = new ODataAdapter('123', store);
+    builder.selectByProjection(this.exportParams.projectionName, true);
+    let colsOrder = settings.colsOrder.filter(({ hide }) => !hide)
+      .map(column => adapter._getODataAttributeName(modelName, column.propName).replace(/\//g, '.') + '/' + column.name || column.propName)
+      .join();
+    if (sortString) {
+      builder.orderBy(sortString);
+    }
+
+    let limitFunction = this.get('objectlistviewEventsService').getLimitFunction();
+    if (limitFunction) {
+      builder.where(limitFunction);
+    }
+
+    if (this.exportParams.isExportExcel) {
+      builder.ofDataType('blob');
+      let customQueryParams = { colsOrder: colsOrder, exportExcel: this.exportParams.isExportExcel,
+        detSeparateRows: this.exportParams.detSeparateRows, detSeparateCols: this.exportParams.detSeparateCols };
+      builder.withCustomParams(customQueryParams);
+    }
+
+    let query = builder.build();
+
+    return query;
   },
 
   _getSavePromise: function(settingName, colsConfig) {
-    return this.get('userSettingsService').saveUserSetting(this.componentName, settingName, colsConfig)
+    return this.get('userSettingsService').saveUserSetting(this.componentName, settingName, colsConfig, this.exportParams.isExportExcel)
     .then(result => {
       this.get('colsConfigMenu').updateNamedSettingTrigger();
     });
@@ -413,7 +554,7 @@ export default FlexberryBaseComponent.extend({
       let tr = trs[i];
       let index = this._getIndexFromId(tr.id);  // get index of initial (model) order
       let colDesc = this.model.colDescs[index];  // Model for this tr
-      colsOrder[i] = { propName: colDesc.propName, hide: colDesc.hide };  //Set colsOrder element
+      colsOrder[i] = { propName: colDesc.propName, hide: colDesc.hide, name: colDesc.name.toString() };  //Set colsOrder element
       if (colDesc.sortPriority !== undefined) { // Sort priority defined
         sortSettings[sortSettings.length] = { propName: colDesc.propName, sortOrder: colDesc.sortOrder, sortPriority: colDesc.sortPriority }; //Add sortSetting element
       }
@@ -444,6 +585,11 @@ export default FlexberryBaseComponent.extend({
     colsConfig = { colsOrder: colsOrder, sorting: sorting, perPage: perPage };  // Set colsConfig Object
     if (this.saveColWidthState) {
       colsConfig.columnWidths = widthSetting;
+    }
+
+    if (this.exportParams.isExportExcel) {
+      colsConfig.detSeparateRows = this.exportParams.detSeparateRows;
+      colsConfig.detSeparateCols = this.exportParams.detSeparateCols;
     }
 
     return colsConfig;