+++ blueprints/flexberry-core/files/__root__/services/store.js
@@ -1,4 +1,13 @@
-import DS from 'ember-data';
-import { Projection } from 'ember-flexberry-data';
+import { Projection, Offline } from 'ember-flexberry-data';
+import config from '../config/environment';
 
-export default DS.Store.reopen(Projection.StoreMixin);
+export default Offline.Store.reopen(Projection.StoreMixin, {
+  init() {
+    this.set('offlineSchema', {
+      [config.APP.offline.dbName]: {
+        // Paste your offline scheme here.
+      }
+    });
+    return this._super(...arguments);
+  }
+});