+++ blueprints/flexberry-model-init/index.js
@@ -6,6 +6,9 @@ module.exports = {
         { name: 'file', type: String },
         { name: 'metadata-dir', type: String }
     ],
+    supportsAddon: function () {
+        return false;
+    },
     /**
      * Blueprint Hook locals.
      * Use locals to add custom template variables. The method receives one argument: options.
@@ -22,6 +25,7 @@ module.exports = {
             className: modelBlueprint.className,
             parentModelName: modelBlueprint.parentModelName,
             parentClassName: modelBlueprint.parentClassName,
+            parentExternal: modelBlueprint.parentExternal,
             name: modelBlueprint.name,
             projections: modelBlueprint.projections,
         };