+++ blueprints/flexberry-model/files/tests/unit/serializers/__test__.js
@@ -6,8 +6,11 @@ moduleForModel('<%= name %>', 'Unit | Serializer | <%= name %>', {
     'serializer:<%= name %>',
     'transform:file',
     'transform:decimal',
+    'transform:guid',
 <% if (!!needsAllEnums === true) { %>
 <%= needsAllEnums %>,
+<% } %><% if (!!needsAllObjects === true) { %>
+<%= needsAllObjects %>,
 <% } %>
 <%= needsAllModels %>
   ]