+++ tests/dummy/app/controllers/components-examples/flexberry-toggler/ge-into-toggler-example.js
@@ -1,0 +1,12 @@
+import ListFormController from 'ember-flexberry/controllers/list-form';
+
+export default ListFormController.extend({
+  /**
+    Name of related edit form route.
+
+    @property editFormRoute
+    @type String
+    @default 'ember-flexberry-dummy-toggler-example-master-e'
+  */
+  editFormRoute: 'ember-flexberry-dummy-toggler-example-master-e',
+});