+++ tests/dummy/app/routes/components-examples/flexberry-objectlistview/downloading-files-from-olv-list.js
@@ -1,0 +1,55 @@
+import ListFormRoute from 'ember-flexberry/routes/list-form';
+
+export default ListFormRoute.extend({
+  /**
+    Name of model projection to be used as record's properties limitation.
+
+    @property modelProjection
+    @type String
+    @default 'SuggestionFileE'
+   */
+  modelProjection: 'SuggestionFileE',
+
+  /**
+  developerUserSettings.
+  {
+  <componentName>: {
+    <settingName>: {
+        colsOrder: [ { propName :<colName>, hide: true|false }, ... ],
+        sorting: [{ propName: <colName>, direction: "asc"|"desc" }, ... ],
+        colsWidths: [ <colName>:<colWidth>, ... ],
+      },
+      ...
+    },
+    ...
+  }
+  For default userSetting use empty name ('').
+  <componentName> may contain any of properties: colsOrder, sorting, colsWidth or being empty.
+
+  @property developerUserSettings
+  @type Object
+  @default {}
+  */
+  developerUserSettings: { DownloadingFilesFromOLV: { } },
+
+  /**
+    Name of model to be used as list's records types.
+
+    @property modelName
+    @type String
+    @default 'ember-flexberry-dummy-suggestion-file'
+   */
+  modelName: 'ember-flexberry-dummy-suggestion-file',
+
+  actions: {
+    objectListViewRowClick(record, params) {
+      // Prevent transition to edit form if cell containing 'flexberry-file' component has been clicked,
+      // and click target isn't cell itself (so target is some element of 'flexberry-file' component).
+      if (params.column && params.column.cellComponent.componentName === 'flexberry-file' && params.originalEvent.target.tagName.toLowerCase() !== 'td') {
+        params.goToEditForm = false;
+      }
+
+      this._super(...arguments);
+    }
+  }
+});