+++ tests/dummy/app/routes/ember-flexberry-dummy-suggestion-list.js
@@ -1,6 +1,7 @@
 import ListFormRoute from 'ember-flexberry/routes/list-form';
+import ListFormRouteOperationsIndicationMixin from '../mixins/list-form-route-operations-indication';
 
-export default ListFormRoute.extend({
+export default ListFormRoute.extend(ListFormRouteOperationsIndicationMixin, {
   /**
    Name of model projection to be used as record's properties limitation.
 
@@ -32,7 +33,11 @@ export default ListFormRoute.extend({
   @default {}
   */
   developerUserSettings: {
-    SuggestionObjectListView: { }
+    SuggestionObjectListView: {
+      'DEFAULT': {
+        'columnWidths': [{ 'propName': 'OlvRowToolbar', 'fixed': true, 'width': 120 }, { 'propName': 'OlvRowMenu', 'fixed': true, 'width': 68 }]
+      }
+    }
   },
 
   /**